# services/chat_service/priv/static/swagger.json
{
  "info": {
    "version": "1.0",
    "description": "Documentation for Chat Service from AshChat Application",
    "title": "AshChat - Chat Service"
  },
...
  "paths": {
...
    "/api/user": {
      "get": {
        "description": "Retrieves the profile of the authenticated user.",
        "operationId": "ChatServiceWeb.UserController.get_user_by_id",
        "parameters": [
          {
            "description": "JWT access token",
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "description": "Device Unique Token",
            "in": "header",
            "name": "device_token",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          [
            "application/json"
          ]
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved",
            "schema": {
              "example": {
                "user": {
                  "description": "Hey, I'm a Fullstack developer!",
                  "nickname": "JohnDoe",
                  "photo_url": "https://static.victor-palha.com/profile.jpg",
                  "preferred_language": "en",
                  "tag_user_id": "ash_chat_17280273"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "example": {
                "error": "User not found"
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          },
          {
            "deviceTokenAuth": []
          }
        ],
        "summary": "Get user profile",
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/user/nickname": {
      "patch": {
        "consumes": [
          [
            "application/json"
          ]
        ],
        "description": "Updates the nickname of the authenticated user.",
        "operationId": "ChatServiceWeb.UserController.update_user_name",
        "parameters": [
          {
            "description": "JWT access token",
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string"
          },
          {
            "description": "Device Unique Token",
            "in": "header",
            "name": "device_token",
            "required": true,
            "type": "string"
          },
          {
            "description": "New nickname",
            "in": "body",
            "name": "nickname",
            "required": true,
            "schema": "string",
            "x-example": "John Doe"
          }
        ],
        "produces": [
          [
            "application/json"
          ]
        ],
        "responses": {
          "200": {
            "description": "User updated successfully",
            "schema": {
              "example": {
                "message": "User updated successfully"
              }
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "example": {
                "error": "User not found"
              }
            }
          }
        },
        "security": [
          {
            "jwtAuth": []
          },
          {
            "deviceTokenAuth": []
          }
        ],
        "summary": "Update user nickname",
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "swagger": "2.0"
}